daemon/config: move "common" tests
